Moving bake_unified to test_utils.py
